+++ migrations/00002-uuid-ossp_to_pgcrypto.sql
@@ -1,0 +1,10 @@
+BEGIN;
+CREATE EXTENSION IF NOT EXISTS pgcrypto;
+
+ALTER TABLE "users" ALTER COLUMN id SET DEFAULT gen_random_uuid();
+ALTER TABLE "attempts" ALTER COLUMN id SET DEFAULT gen_random_uuid();
+ALTER TABLE "comments" ALTER COLUMN id SET DEFAULT gen_random_uuid();
+ALTER TABLE "reactions" ALTER COLUMN id SET DEFAULT gen_random_uuid();
+
+DROP EXTENSION IF EXISTS "uuid-ossp";
+COMMIT;