+++ scripts/seed/users.sql
@@ -1,19 +1,13 @@
-/* @name GetUserById */
-SELECT
-  *
-FROM
-  users
-WHERE
-  id = :userId;
-
-
-/* 
- @name InsertTestUser
- @param user -> (firstName, lastName, email, password, username)
+/*
+ @name GetOrCreateUser
+ @param user -> (id, firstName, lastName, email, password, username)
  */
-INSERT INTO users (first_name, last_name, email, "password", username)
+INSERT INTO users (id, first_name, last_name, email, "password", username)
   VALUES
     :user
-  RETURNING
-    *;
+  ON CONFLICT ON CONSTRAINT "pk_users_id"
+    DO UPDATE SET
+      id = users.id
+    RETURNING
+      *;