+++ src/resolvers/Mutation/createComment/index.ts
@@ -1,0 +1,1 @@
+export { createComment } from "./createComment"