+++ src/resolvers/Mutation/requestPasswordResetLink.ts
@@ -3,8 +3,8 @@ import ms from "ms"
 import { SESV2 } from "aws-sdk"
 
 import { ResolverFn } from "resolvers"
-import { APP_SECRET } from "../../utils"
-import { User } from "../../entity/User"
+import { APP_SECRET } from "utils"
+import { User } from "entity/User"
 import { createPasswordResetEmailHTMLString } from "../password_reset_email"
 
 export type RequestPasswordResetLinkArgs = {