+++ src/resolvers/Mutation/updatePassword/updatePassword.sql
@@ -1,0 +1,11 @@
+/*
+ @name getUserById
+ */
+SELECT * FROM users WHERE id = :userId; 
+
+/*
+ @name updatePasswordForUserId
+ */
+UPDATE users
+SET "password" = :newHash
+WHERE id = :userId;