+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2020-08-06
+
+### Added
+
+- Added missing enum to `00001-Bootstrap.sql` migration
+- Enabled absolute imports ✨
+- Added `@auth` and `@development` GraphQL directives
+
+### Changed
+
+- Refactored resolvers to use PgTyped
+  - `updatePassword`
+  - `createComment`
+  - `signup`
+
 ## [0.1.0] - 2020-07-26
 
 ### Added
@@ -95,6 +110,7 @@ add skip/take to getCommentsByUrl
 
 ### Added
 
+[0.1.1]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.0.10...v0.1.0
 [0.0.10]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.0.8...v0.0.9